project/requirements.txt
Flask
playwright
beautifulsoup4
requests
python-dotenv
groq
geopy
aiohttp
tenacity
